# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Cadastro de Clientes - Escritório Mendes Advocacia
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Cliente {
  id: number;
  escritorio: string;
  cpf: string;
  nomeCompleto: string;
  /** @nullable */
  dataNascimento?: string | null;
  /** @nullable */
  sexo?: string | null;
  /** @nullable */
  estadoCivil?: string | null;
  /** @nullable */
  rgRepresentante?: string | null;
  /** @nullable */
  orgaoEmissor?: string | null;
  /** @nullable */
  profissao?: string | null;
  /** @nullable */
  telefone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  cep?: string | null;
  /** @nullable */
  estado?: string | null;
  /** @nullable */
  cidade?: string | null;
  /** @nullable */
  logradouro?: string | null;
  /** @nullable */
  numero?: string | null;
  /** @nullable */
  complemento?: string | null;
  /** @nullable */
  bairro?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  pastaPath?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClienteInput {
  escritorio: string;
  cpf: string;
  nomeCompleto: string;
  /** @nullable */
  dataNascimento?: string | null;
  /** @nullable */
  sexo?: string | null;
  /** @nullable */
  estadoCivil?: string | null;
  /** @nullable */
  rgRepresentante?: string | null;
  /** @nullable */
  orgaoEmissor?: string | null;
  /** @nullable */
  profissao?: string | null;
  /** @nullable */
  telefone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  cep?: string | null;
  /** @nullable */
  estado?: string | null;
  /** @nullable */
  cidade?: string | null;
  /** @nullable */
  logradouro?: string | null;
  /** @nullable */
  numero?: string | null;
  /** @nullable */
  complemento?: string | null;
  /** @nullable */
  bairro?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  pastaPath?: string | null;
}

export interface Processo {
  id: number;
  clienteId: number;
  /** @nullable */
  numero?: string | null;
  /** @nullable */
  vara?: string | null;
  /** @nullable */
  comarca?: string | null;
  /** @nullable */
  assunto?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProcessoInput {
  /** @nullable */
  numero?: string | null;
  /** @nullable */
  vara?: string | null;
  /** @nullable */
  comarca?: string | null;
  /** @nullable */
  assunto?: string | null;
  status: string;
}

export interface Anexo {
  id: number;
  clienteId: number;
  tipo: string;
  nomeArquivo: string;
  /** @nullable */
  fileData?: string | null;
  createdAt: string;
}

export interface CreateAnexoInput {
  tipo: string;
  nomeArquivo: string;
  /** @nullable */
  fileData?: string | null;
}

export type ListClientesParams = {
  /**
   * Busca por CPF ou nome
   */
  search?: string;
};