# lib/api-zod/src/generated/types/anexo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Cadastro de Clientes - Escritório Mendes Advocacia
 * OpenAPI spec version: 0.1.0
 */

export interface Anexo {
  id: number;
  clienteId: number;
  tipo: string;
  nomeArquivo: string;
  /** @nullable */
  fileData?: string | null;
  createdAt: string;
}