# lib/api-zod/src/generated/types/createClienteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Cadastro de Clientes - Escritório Mendes Advocacia
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClienteInput {
  escritorio: string;
  cpf: string;
  nomeCompleto: string;
  /** @nullable */
  dataNascimento?: string | null;
  /** @nullable */
  sexo?: string | null;
  /** @nullable */
  estadoCivil?: string | null;
  /** @nullable */
  rgRepresentante?: string | null;
  /** @nullable */
  orgaoEmissor?: string | null;
  /** @nullable */
  profissao?: string | null;
  /** @nullable */
  telefone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  cep?: string | null;
  /** @nullable */
  estado?: string | null;
  /** @nullable */
  cidade?: string | null;
  /** @nullable */
  logradouro?: string | null;
  /** @nullable */
  numero?: string | null;
  /** @nullable */
  complemento?: string | null;
  /** @nullable */
  bairro?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  pastaPath?: string | null;
}