# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Cadastro de Clientes - Escritório Mendes Advocacia
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}